Fix test broken by Java 16 deprecation fixes

+++ test/org/apache/juli/TestThreadNameCache.java
@@ -32,7 +32,7 @@ public class TestThreadNameCache {
         final CountDownLatch cacheLatch = new CountDownLatch(1);
 
         OneLineFormatter olf = new OneLineFormatter();
-        Method getThreadName = olf.getClass().getDeclaredMethod("getThreadName", int.class);
+        Method getThreadName = olf.getClass().getDeclaredMethod("getThreadName", long.class);
         getThreadName.setAccessible(true);
         Thread thread = new Thread() {
             @Override
